ajout de règle exception sur date debut date fin, remboursement ok logiquement.

+++ src/main/java/fr/eni/encheres/repository/ArticleRepository.java
@@ -37,5 +37,5 @@ public interface ArticleRepository {
 
     int updateECtoVD();
 
-    long finalPointInProgress(long id_article);
+    int finalPointInProgress(int id_article);
 }
